feat: changed route layout to accomodate for header and footer
