+++ src/app/(main)/article/[slug]/page.tsx
@@ -1,0 +1,9 @@
+type ArticlePageProps = {
+  params: Promise<{ slug: string }>
+}
+
+export default async function ArticlePage({ params }: ArticlePageProps) {
+  const { slug } = await params
+
+  return <div>Article</div>
+}
