--- src/app/article/[slug]/page.tsx
@@ -1,11 +1,0 @@
-type ArticlePageProps = {
-  params: Promise<{ slug: string }>;
-};
-
-// TODO: Bring article details
-
-export default async function ArticlePage({ params }: ArticlePageProps) {
-  const { slug } = await params;
-
-  return <div>Article</div>;
-}
